+++ src/Attribute/Constraint/UniqueField.php
@@ -15,6 +15,7 @@ class UniqueField extends Constraint
         public readonly string $message = 'unique_field.not_unique',
         public readonly string|null $em = null,
         public readonly string|null $ignoreProperty = null,
+        public readonly string|null $ignoreRouteParam = null,
         array|null $groups = null,
         mixed $payload = null,
     ) {